requirements.txt
cycler==0.10.0
kiwisolver==1.3.1
matplotlib==3.4.2
numpy==1.20.3
pandas==1.2.4
Pillow==8.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
scipy==1.6.3
seaborn==0.11.1
selenium==3.141.0
six==1.16.0
urllib3==1.26.5
